+++ standalone/GripperHmiTool/include/gripper_hmi_led_effects.h
@@ -8,7 +8,9 @@
 
 enum class GripperLedEffectState
 {
+    BootInit,
     Init,
+    Writing,
     WaitStorage,
     Ready,
     Warning,